Обновил модели, контроллеры, сервисы, фасад

+++ src/main/java/com/itstep/travelpackage/controller/BaseController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.*;
 public interface BaseController<D, DC, DU> {
 
     @GetMapping("/one")
-    ResponseEntity<D> findOne(@RequestParam Integer id);
+    ResponseEntity<D> findOne(@RequestParam Long id);
 
     @PostMapping("/create")
     ResponseEntity<D> create(@RequestBody DC dto);
@@ -15,5 +15,5 @@ public interface BaseController<D, DC, DU> {
     ResponseEntity<D> update(@RequestBody DU dto);
 
     @DeleteMapping("/remove")
-    void remove(@RequestParam Integer id);
+    void remove(@RequestParam Long id);
 }
