+++ src/main/java/com/itstep/travelpackage/repository/TravelFeedRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TravelFeedRepository extends JpaRepository<TravelFeed, Integer>, JpaSpecificationExecutor<TravelFeed> {
+public interface TravelFeedRepository extends JpaRepository<TravelFeed, Long>, JpaSpecificationExecutor<TravelFeed> {
 }
